% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcrel.R
\name{calcrel}
\alias{calcrel}
\title{Calculation of reliability (calcrel)}
\usage{
calcrel(d1,d2)
}
\arguments{
\item{d1}{list of numbers from measurement one}

\item{d2}{list of numbers from measurement two (same order as for measurement one)}
}
\value{
Returns a nested list of reliability measures.
}
\description{
\code{calcrel()} is a function used to calculate different reliability measures, including Coefficient of variance, smallest real difference, intraclass correlation coefficient, and Bland-Altman plot derived bias with 95\% limits of agreement.
}
\examples{
   d1 <- rnorm(15,10,1)
   d2 <- rnorm(15,10,1)
   calcrel(d1,d2)

}
