% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danishcpr.R
\name{danishcpr}
\alias{danishcpr}
\title{Danish CPR number to birthday and sex (danishcpr)}
\usage{
danishcpr(code)
}
\arguments{
\item{code}{list of CPR numbers (\code{list})}
}
\value{
Returns a list with $birthday which is a list of dates, and $sex which is a list of Male/Female.
}
\description{
\code{danishcpr()} converts a list of CPR-numbers to a corresponding list birthday and sex.
}
\examples{
\dontrun{
   cpr <- danishcpr(code)
   birthdays <- cpr$birthday
   sex <- cpr$sex
}

}
