% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_testdata.R
\docType{data}
\name{data10}
\alias{data10}
\title{Test-data (10 Hz)}
\format{
An object of class \code{"dataframe"}; an example of the
usage in \code{\link[clinmon]{Mx}}.
}
\source{
\href{https://github.com/lilleoel/clinmon}{GitHub}
}
\usage{
data(testdata)
}
\description{
Recording with four columns: time (t), non-invasive arterial
blood pressure (nabp), middle cerebral artery velocity measured
using transcranial Doppler (tcd), and heart rate (hr).
}
\examples{
data(testdata)
Mx(data10[,c(1:3)],freq=10)
}
\references{
Olsen MH et al. (Unpublished data, 2020)
(\href{https://github.com/lilleoel/clinmon}{GitHub})
}
\keyword{datasets}
