% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_testdata.R
\docType{data}
\name{del}
\alias{del}
\title{Test-deleter}
\format{
An object of class \code{"dataframe"}; an example of the
usage in \code{\link[clinmon]{Mx}}.
}
\source{
\href{https://github.com/lilleoel/clinmon}{GitHub}
}
\usage{
data(testdata)
}
\description{
Deleter dataframe with two columns: start (start) and
end (end) of the deleter-period.
}
\examples{
data(testdata)
Mx(data1000[,c(1:3)],del=del,freq=1000)
}
\references{
Olsen MH et al. (Unpublished data, 2020)
(\href{https://github.com/lilleoel/clinmon}{GitHub})
}
\keyword{datasets}
