% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.data10}
\alias{df.data10}
\title{Test-data (10 Hz)}
\format{
An object of class \code{"dataframe"}; an example of the
usage in \code{\link[clintools]{clinmon}}-function.
}
\source{
\href{https://github.com/lilleoel/clintools}{GitHub}
}
\usage{
data(testdata)
}
\description{
Recording with four columns: time (\code{t}), non-invasive arterial
blood pressure (\code{abp}), middle cerebral artery velocity measured
using transcranial Doppler (\code{mcav}), and heart rate (\code{hr}).
}
\examples{
data(testdata)
variables <- c("abp","mcav","hr")
clinmon(df.data10,variables,freq=10)
}
\references{
Olsen MH et al. (Unpublished data, 2020) (\href{https://github.com/lilleoel/clintools}{GitHub})
}
\keyword{datasets}
