% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{df.deleter}
\alias{df.deleter}
\title{Test-deleter}
\format{
An object of class \code{"dataframe"}; an example of the
usage in \code{\link[clinmon]{clinmon}}.
}
\source{
\href{https://github.com/lilleoel/clinmon}{GitHub}
}
\usage{
data(testdata)
}
\description{
Deleter dataframe with two columns: start (start) and
end (end) of the deleter-period.
}
\references{
Olsen MH et al. (Unpublished data, 2020) (\href{https://github.com/lilleoel/clinmon}{GitHub})
}
\keyword{datasets}
