% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscus.R
\name{iscus}
\alias{iscus}
\title{ISCUSFlex-values to dataframe  (iscus)}
\usage{
iscus(filename)
}
\arguments{
\item{filename}{path to the XML-file with the measurements}
}
\value{
Returns a dataframe with the measurements.
}
\description{
\code{iscus()} is a function which converts XML files extracted
from the Microdialyis-apparatur of ISCUSFlex apparatus to a dataframe.
}
\examples{
 \dontrun{
   iscus("C:/ISCUSfiles/7888e844-1c7a-40af-a3f2-3bb27a8dd9e5.xml")
 }

}
