% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{libraries}
\alias{libraries}
\title{Load or install multiple packages (libraries)}
\usage{
libraries(packages)
}
\arguments{
\item{packages}{list of packages}
}
\value{
outputs the loaded and/or installed packages
}
\description{
\code{libraries()} is a small function which can load and/or install multiple packages
}
\examples{
\dontrun{
libraries(c("rjson", "data.table"))
}

}
