% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metsz.R
\name{metsz}
\alias{metsz}
\title{Calculation of MetS-Z score (metsz)}
\usage{
metsz(d, id, ethnicity, sex, wc, hdl, sbp, tri, glu)
}
\arguments{
\item{d}{the dataframe with the data}

\item{id}{colname of the id column}

\item{ethnicity}{colname of the ethnicity column}

\item{sex}{colname of the sex column (must be labelled Male and Female)}

\item{wc}{colname of the waist circumference}

\item{hdl}{colname of HDL}

\item{sbp}{colname of systolic blood pressure}

\item{tri}{colname of Triglyceride}

\item{glu}{colname of Glucose}
}
\value{
Returns a dataframe with ssid and MetS-Z score
}
\description{
\code{metsz()} is a small function which calculate the MetS-Z score from danish SI values.
}
\examples{
\dontrun{
mm <- metsz(df,id="ssid",ethnicity = "SCRE.SC05_ethnicity",
sex = "Sex", wc = paste0(x,".wc"),
hdl = paste0(x,".BR15_cholesterol_hdl_p"),
sbp = paste0(x,".sbp"), tri = paste0(x,".BR17_triglycerid_p"),
glu = paste0(x,".BR01_glucose_p"))
}
}
