% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparacontcomp.R
\name{nonparacontcomp}
\alias{nonparacontcomp}
\title{Non-parametric Comparison of Variable Columns}
\usage{
nonparacontcomp(working_data, var1, var2)
}
\arguments{
\item{working_data}{data frame to work with}

\item{var1}{var1 for comparison}

\item{var2}{var2 for comparison}
}
\description{
Gets the median, calculates the IQR and gets the p-value using Fisher's Exact test
}
\examples{
nonparacontcomp()
}
