% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.col.missing.R
\name{percent.col.missing}
\alias{percent.col.missing}
\title{Percent missing of each column of data frame}
\usage{
percent.col.missing(df)
}
\arguments{
\item{df}{dataframe to analyze}
}
\description{
Takes in a dataframe and gives you what percent of each row is missing 
Outputs missingess of each column
}
\examples{
percent.col.missing(df) # just pass it a dataframe
percent.col.missing()
}
\keyword{blank}
