% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLR.R
\name{PLR}
\alias{PLR}
\title{NeurOpticsTM PLR-3000 or PLR-4000 pupillometer file to dataframe (PLR)}
\usage{
PLR(filename = NULL, df = NULL, v = "3000")
}
\arguments{
\item{filename}{path to the XLS-file with the measurements}

\item{df}{the dataframe can also be used for the function if data is already imported.}

\item{v}{needed to define if it is either "3000" or "4000"}
}
\value{
Returns a list with two dataframe, one with the measurements (pupils) and one with the markers (markers).
}
\description{
\code{PLR()} is a function which converts the XLS file imported from the NeurOpticsTM PLR-3000 or PLR-4000 pupillometer to a nested list with two dataframes.
}
\examples{
 \dontrun{
   PLR0("C:/PLR3000/R_20200105_205901.xls")
 }

}
