% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionaire.R
\name{questionaire}
\alias{questionaire}
\title{Calculate scores from questionaire (questionaire)}
\usage{
questionaire(df,id,questions,scale,setting,prefix)
}
\arguments{
\item{df}{dataframe. (\code{df})}

\item{id}{Column name of participant id (\code{string})}

\item{questions}{Column names of ordered list of questions (\code{list})}

\item{scale}{name of the questionaire (\code{string})}

\item{setting}{questionaire specific settings (\code{string})}

\item{prefix}{prefix of column names of questionaire scores (\code{string})}
}
\value{
Returns summarised information in dataframe.
}
\description{
\code{questionaire()} is a function which calculates scores from a questionaire.
}
\examples{
\dontrun{

   df <- df_b[,grepl("ssid|KIDSCREEN52_D_",colnames(df_b))][,c(1:53)]
   k52 <- questionaire(df,id = "ssid",
   questions=colnames(df)[grepl("KIDSCREEN",colnames(df))],
   scale="Kidscreen-52",
   setting="proxy")#'
}

}
