% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quietly.R
\name{quietly}
\alias{quietly}
\title{Quiet any output (quietly)}
\usage{
quietly(x)
}
\arguments{
\item{x}{input to be suppressed}
}
\value{
Returns x, but without any output
}
\description{
\code{quietly()} is a small function which suppresses any output
}
\examples{
tmp <- quietly(print("hello"))

}
