% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.openclinica.R
\name{read.openclinica}
\alias{read.openclinica}
\title{Create dataframe from Open Clinica (read,openclinica)}
\usage{
read.openclinica(trial, link, prefix, ids)
}
\arguments{
\item{trial}{The trialname used in openclinica for extraction}

\item{link}{The link where to access data}

\item{prefix}{for each instrument how many characters should be used in the initial part of the colnames}

\item{ids}{list with ID's which reccur from each instrument}
}
\value{
Returns a dataframe with data from Open Clinica
}
\description{
\code{read.openclinica()} is a small function which can extract data from Open Clinica
}
\examples{
\dontrun{
df <- read.openclinica(trial="LightCARE",link = "https://lightcom.ctu.dk/extract/api/data/",
ids = c("ssid","site_id"))
}
}
