% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savenprint.R
\name{savenprint}
\alias{savenprint}
\title{Save and print (savenprint)}
\usage{
savenprint(plot, height, width, dpi, filename, folder, print)
}
\arguments{
\item{plot}{input a ggplot}

\item{height}{height in inches}

\item{width}{width in inches}

\item{dpi}{prefer 600 dpi to print (which is default)}

\item{filename}{the filename - if null then automatically generated}

\item{folder}{the folder where the figures whould be stored}

\item{print}{boolean to say if plot will be printed}
}
\value{
prints and save a figure
}
\description{
\code{savenprint()} is a small function which saves figures as *.jpg and prints the figure in a markdown document.
}
\examples{
\dontrun{
p <- ggplot(dfa,aes(x=Var2,y=prop,color=Var1)) +
geom_vline(aes(xintercept=98),linetype="dotted") +
geom_line() + geom_point() +
theme_classic()

savenprint(p, height = 4, width = 8,
   filename = "Figure1", folder="figures")
}

}
