% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup.R
\name{subgroups}
\alias{subgroups}
\title{Create subgroup analysis table (subgroups)}
\usage{
subgroups(df,group,subgroups,outcome,test,time.to,
                  strata.fixed,strata.random,digs_p,digs_s,
                  conf.level, paired,markdown, caption)
}
\arguments{
\item{df}{dataframe. (\code{df})}

\item{group}{Column name of stratification (\code{string})}

\item{subgroups}{column names of subgroups (\code{list})}

\item{outcome}{column name of outcome of interest (\code{string})}

\item{test}{list of tests carried out, currently the following works: \code{t.test}, \code{wilcox.test}, \code{fisher.test}, \code{auc}, \code{lm}, and \code{glm}. (\code{list})}

\item{time.to}{Column name of the time column for cox regression (\code{list})}

\item{strata.fixed}{list of columns which should be used as fixed stratification (\code{list})}

\item{strata.random}{list of columns which should be used as random stratification (\code{list})}

\item{digs_p}{digits for p-values (\code{numeric})}

\item{digs_s}{digits for statistics (\code{numeric})}

\item{conf.level}{confidence intervals which should be presented (\code{numeric}).}

\item{paired}{if tests should be paired (\code{boolean})}

\item{markdown}{default is true and output is pander, while false output is a dataframe (\code{boolean})}

\item{caption}{Table caption only in use when markdown is true (\code{string})}
}
\value{
Returns summarised information in dataframe.
}
\description{
\code{subgroups()} is a function which create a dataframe, which can be copied directly into
word or presented in as a subgroup analysis.
}
\examples{
\dontrun{
   zubgroups <- c(`Estimated fetal weight`="est_fetal_weight_grp",
   `Gestational age`="Gestational Age at rand (cat)",
   `End diastolic flow`="end_dia_flow_grp",
   `Uterine artery Doppler / Notch`="dop_notch_grp",
   `Maternal S-Placental-GF`="maternalSplacentalGF_grp",
   `Hypertension or preeclampsia`="ht_preeclamps_grp",
   `Abdominal circumference`="abd_circum_cent_grp")

   subgroups(df,group="group",zubgroups,
      outcome="Infant survival without SAE",
      test="glm",strata.fixed = "site")
}

}
