% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{df.data10}
\alias{df.data10}
\title{Test-data (10 Hz)}
\format{
An object of class \code{"dataframe"}; an example of the
usage in \code{\link[clinmon]{clinmon}}.
}
\source{
\href{https://github.com/lilleoel/clinmon}{GitHub}
}
\usage{
data(testdata)
}
\description{
Recording with four columns: time (t), non-invasive arterial
blood pressure (abp), middle cerebral artery velocity measured
using transcranial Doppler (mcav), and heart rate (hr).
}
\references{
Olsen MH et al. (Unpublished data, 2020) (\href{https://github.com/lilleoel/clinmon}{GitHub})
}
\keyword{datasets}
