% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullxtab.R
\name{fullxtab}
\alias{fullxtab}
\title{Improved xtab}
\usage{
fullxtab(var1, var2)
}
\arguments{
\item{var1}{rows for the xtab}
}
\description{
Gets the xtab, calculates the percentages and gets the p-value
}
\examples{
fullxtab_function(var1, var2)
fullxtab_function()
}
\keyword{blank}
