% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullxtab.R
\name{fullxtab}
\alias{fullxtab}
\title{Improved xtab}
\usage{
fullxtab(working_data, var1, var2)
}
\arguments{
\item{working_data}{data frame to work with}

\item{var1}{colname as a string, the categories of this var will be the row names for the xtab}

\item{var2}{colname as a string, will the cols for the xtab, needs to be in 0 / 1 form}
}
\description{
Gets the xtab, calculates the percentages and gets the p-value
}
\examples{
fullxtab_function(var1, var2)
fullxtab_function()
}
\keyword{blank}
