% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparacontcomp.R
\name{nonparacontcomp}
\alias{nonparacontcomp}
\title{Non-parametric Comparison of Variable Columns}
\usage{
nonparacontcomp(var1, var2)
}
\arguments{
\item{var1}{var1 for comparison}
}
\description{
Gets the median, calculates the IQR and gets the p-value using Fisher's Exact test
}
\examples{
nonparacontcomp()
}
\keyword{comparison}
\keyword{for}
\keyword{var2}
